Spell file restructure
