+++ game/world/managers/objects/UnitManager.py
@@ -6,8 +6,8 @@ from database.dbc.DbcDatabaseManager import DbcDatabaseManager
 from game.world.managers.GridManager import GridManager
 from game.world.managers.objects.MovementManager import MovementManager
 from game.world.managers.objects.ObjectManager import ObjectManager
-from game.world.managers.objects.AuraManager import AuraManager
-from game.world.managers.objects.SpellManager import SpellManager
+from game.world.managers.objects.spell.AuraManager import AuraManager
+from game.world.managers.objects.spell.SpellManager import SpellManager
 from network.packet.PacketWriter import PacketWriter, OpCode
 from network.packet.update.UpdatePacketFactory import UpdatePacketFactory
 from utils import Formulas
